添加卷轴、矿石、椅子、坐骑背包功能

+++ gms-server/src/main/java/org/gms/server/StorageInventory.java
@@ -33,11 +33,11 @@ import java.util.*;
 public class StorageInventory {
     private final Client c;
     private Map<Short, Item> inventory = new LinkedHashMap<>();
-    private final byte slotLimit;
+    private final int slotLimit;
 
     public StorageInventory(Client c, List<Item> toSort) {
         this.inventory = new LinkedHashMap<>();
-        this.slotLimit = (byte) toSort.size();
+        this.slotLimit = toSort.size();
         this.c = c;
 
         for (Item item : toSort) {
@@ -45,7 +45,7 @@ public class StorageInventory {
         }
     }
 
-    private byte getSlotLimit() {
+    private int getSlotLimit() {
         return slotLimit;
     }
